backend/pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Gregory Glasby",email = "gglasby@unc.edu"}
]
requires-python = ">=3.9"
dependencies = [
    "djangorestframework (>=3.16.1,<4.0.0)",
    "markdown (>=3.9,<4.0)",
    "pyyaml (>=6.0.3,<7.0.0)",
    "requests (>=2.32.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
